+++ local_image_provider/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 4.2.0
+* Android method signature change for Flutter 3.0.0
+* Fix for reference to non-null PaintingBinding.instance in [DeviceImage]
+
+## 4.1.0
+* Fix for image bytes write [#42](https://github.com/csdcorp/local_image_provider/issues/42)
+
 ## 4.1.0
 * Required dependencies now exported, no longer need explicit import 
 for `DeviceImage`, `LocalImage`, `LocalAlbum`